# Implementation pack/05-openapi.yaml
openapi: 3.0.3
info:
  title: AI CV Tailoring and Cover Letter Platform — Backend API
  description: >
    Draft specification for developer implementation. Covers CV ingestion,
    job post ingestion, matching, tailored CV generation, guided cover
    letter generation, and export. All generated content must carry
    evidence references — see evidenceReferences fields throughout.
  version: 0.1.0-draft
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /exports/cv/{draftId}:
    parameters:
      - $ref: '#/components/parameters/draftId'
    post:
      tags: [exports]
      summary: Request a docx export package for an approved CV draft
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                templateId: { type: string, description: "CV template layout id — see GET /exports/templates. Defaults to 'standard' if omitted or unrecognized." }
      responses:
        '202':
          description: Export queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportRequest' }
        '409':
          description: Draft is not approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /exports/cover-letter/{workflowId}:
    parameters:
      - $ref: '#/components/parameters/workflowId'
    post:
      tags: [exports]
      summary: Request a docx export package for an approved cover letter
      responses:
        '202':
          description: Export queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportRequest' }
        '409':
          description: Draft is not approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /exports/application-pack:
    post:
      tags: [exports]
      summary: Export approved CV plus approved cover letter together, as a zip of two independent docx files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tailoredCvDraftId, coverLetterWorkflowId]
              properties:
                tailoredCvDraftId: { type: string, format: uuid }
                coverLetterWorkflowId: { type: string, format: uuid }
                templateId: { type: string, description: "CV template layout id, applied to the CV half of the pack." }
      responses:
        '202':
          description: Export queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportRequest' }
        '409':
          description: One or both drafts are not approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /exports/{exportId}:
    parameters:
      - name: exportId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [exports]
      summary: Poll an export's status; fileReference becomes a download URL once completed
      responses:
        '200':
          description: Export status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /exports/{exportId}/download:
    parameters:
      - name: exportId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [exports]
      summary: Download a completed export's file (proxied, ownership-checked every request — no presigned URL)
      responses:
        '200':
          description: The exported file (docx, pdf, or zip, per the export's format field)
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document: {}
            application/pdf: {}
            application/zip: {}
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Export exists but isn't completed yet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /exports/{exportId}/pdf:
    parameters:
      - name: exportId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [exports]
      summary: Convert an already-downloaded docx export to PDF (secondary format, gated on download having happened first)
      responses:
        '202':
          description: PDF conversion queued — the response is a new Export row (format=pdf, derivedFromExportId set to the source docx export)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: "Source export isn't docx, isn't completed yet, or hasn't been downloaded yet"
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    draftId:
      name: draftId
      in: path
      required: true
      schema: { type: string, format: uuid }
    workflowId:
      name: workflowId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: >
        Standard error envelope returned by every non-2xx response.
        `details` is free-form and endpoint-specific — populate it with
        whatever helps the caller correct the request (offending field,
        allowed values, etc.), omit it when there's nothing further to add.
      required: [status, code, message, timestamp, path]
      properties:
        status: { type: integer, description: "HTTP status code, duplicated in the body for logging convenience" }
        code: { type: string, description: "Stable machine-readable error code, e.g. VALIDATION_ERROR" }
        message: { type: string, description: "Human-readable explanation" }
        timestamp: { type: string, format: date-time }
        path: { type: string, description: "Request path that produced the error" }
        details:
          type: object
          additionalProperties: true
          nullable: true
      example:
        status: 400
        code: VALIDATION_ERROR
        message: "Invalid file type. Only PDF and DOCX are supported."
        timestamp: "2026-08-03T10:00:00Z"
        path: "/api/v1/cvs"
        details:
          field: file
          reason: unsupported_mime_type
          allowed: ["application/pdf", "application/vnd.openxmlformats-officedocument.wordprocessingml.document"]
...
    ExportRequest:
      type: object
      description: >
        Sprint 5. format resolves the format decision this schema originally
        left as a placeholder (see 01-implementation-plan.md open decisions):
        docx (primary), pdf (secondary, only after downloadedAt is set), zip
        (application packs — two independent docx files, not a merged doc).
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [queued, processing, completed, failed] }
        sourceDraftType: { type: string, enum: [tailored_cv, cover_letter, application_pack] }
        fileReference: { type: string, nullable: true, description: "Relative download URL (GET /exports/{id}/download) once status is completed; null otherwise." }
        format: { type: string, enum: [docx, pdf, zip] }
        templateId: { type: string, nullable: true, description: "CV template layout used; null for cover-letter/pdf/zip exports." }
        downloadedAt: { type: string, format: date-time, nullable: true, description: "Set the first time the docx has been downloaded — gates PDF conversion." }
        derivedFromExportId: { type: string, format: uuid, nullable: true, description: "Set on format=pdf rows, pointing at the source docx export." }
        errorMessage: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...